+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/AnnotationSpacingRule.kt
@@ -9,6 +9,7 @@ import com.pinterest.ktlint.core.ast.isWhiteSpaceWithNewline
 import com.pinterest.ktlint.core.ast.nextCodeSibling
 import com.pinterest.ktlint.core.ast.nextLeaf
 import com.pinterest.ktlint.core.ast.nextSibling
+import com.pinterest.ktlint.core.ast.prevSibling
 import org.jetbrains.kotlin.com.intellij.lang.ASTNode
 import org.jetbrains.kotlin.com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
@@ -16,6 +17,7 @@ import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.PsiWhiteSpaceImpl
 import org.jetbrains.kotlin.psi.KtAnnotationEntry
 import org.jetbrains.kotlin.psi.psiUtil.children
 import org.jetbrains.kotlin.psi.psiUtil.endOffset
+import org.jetbrains.kotlin.psi.psiUtil.leaves
 
 /**
  * Ensures annotations occur immediately prior to the annotated construct
@@ -62,7 +64,8 @@ class AnnotationSpacingRule : Rule("annotation-spacing") {
             {
                 !it.isWhiteSpace() &&
                     it.textLength > 0 &&
-                    !it.isPartOf(ElementType.FILE_ANNOTATION_LIST)
+                    !it.isPartOf(ElementType.FILE_ANNOTATION_LIST) &&
+                    !it.isCommentOnSameLineAsPrevLeaf()
             },
             {
                 // Disallow multiple white spaces as well as comments
@@ -70,7 +73,7 @@ class AnnotationSpacingRule : Rule("annotation-spacing") {
                     val s = it.text
                     // Ensure at least one occurrence of two line breaks
                     s.indexOf("\n") != s.lastIndexOf("\n")
-                } else it.isPartOfComment()
+                } else it.isPartOfComment() && !it.isCommentOnSameLineAsPrevLeaf()
             }
         )
         if (next != null) {
@@ -82,13 +85,25 @@ class AnnotationSpacingRule : Rule("annotation-spacing") {
                     // by a comment: we need to swap the order of the comment and the annotation
                     if (next.isPartOfComment()) {
                         // Remove the annotation and the following whitespace
-                        val nextSibling = node.nextSibling { it.isWhiteSpace() }
+                        val eolComment = node.nextSibling { it.isCommentOnSameLineAsPrevLeaf() }
+                        if (eolComment != null) {
+                            eolComment.prevSibling { it.isWhiteSpace() }?.let { it.treeParent.removeChild(it) }
+                            eolComment.nextSibling { it.isWhiteSpace() }?.let { it.treeParent.removeChild(it) }
+                            eolComment.treeParent?.removeChild(eolComment)
+                        } else {
+                            node.nextSibling { it.isWhiteSpace() }?.let { it.treeParent?.removeChild(it) }
+                        }
                         node.treeParent.removeChild(node)
-                        nextSibling?.treeParent?.removeChild(nextSibling)
+
                         // Insert the annotation prior to the annotated construct
-                        val space = PsiWhiteSpaceImpl("\n")
-                        next.treeParent.addChild(space, next.nextCodeSibling())
-                        next.treeParent.addChild(node, space)
+                        val beforeAnchor = next.nextCodeSibling()
+                        val treeParent = next.treeParent
+                        treeParent.addChild(node, beforeAnchor)
+                        if (eolComment != null) {
+                            treeParent.addChild(PsiWhiteSpaceImpl(" "), beforeAnchor)
+                            treeParent.addChild(eolComment, beforeAnchor)
+                        }
+                        treeParent.addChild(PsiWhiteSpaceImpl("\n"), beforeAnchor)
                     } else {
                         removeExtraLineBreaks(node)
                     }
@@ -165,4 +180,7 @@ class AnnotationSpacingRule : Rule("annotation-spacing") {
             removeIntraLineBreaks(lNext, last)
         }
     }
+
+    private fun ASTNode.isCommentOnSameLineAsPrevLeaf() =
+        isPartOfComment() && leaves(forward = false).takeWhile { it.isWhiteSpace() }.none { "\n" in it.text }
 }